Move HTTP API enums to own namespace
